Discriminate values of gcc UInteger options
When an option takes an integral argument, it can be conditionally
enabled at a specific value or be enabled by a value comparison. For
example, -Wall enables -Wformat=1, and -Wformat-y2k is only enabled
by -Wformat=2. Previously the value was not considered, leading to
incorrect implications such as
* -Wall enables -Wformat-y2k
* -Wall enables the strictest level of -Warray-bounds=

This patch adds parsing of the posarg atom of LangEnabledBy to find
the enabled value (in the case of -Wall enabling -Wformat=1) and
enablement comparisons (for the case of -Wformat=2 enabling
-Wformat-y2k). Further, valued forms of options are added to the
complete warning list for values that are seen during such parsing.