gcc: Introduce GccDiagnostics and GccOption
Introduce a GccOption class to associate the data from parsing an
option with a single object, rather than being distributed across
multiple containers.

Introduce a GccDiagnostics class as the container for all
GccOption objects, with helper methods to parse the options file
and perform queries that need the container.

This change also fixes the language processing to merge rather than
replace. In particular this corrects the handling of the
Wdelete-non-virtual-dtor option in gcc 4.8 and 4.9, as those versions
had two records for this option.