fix(clang): Tolerate unparsable format strings
In parse-clang-diagnostic-groups, resolve_format_string() raises a
RuntimeError when the diagnostic message contains an unexpected format
string. This caused the script to crash. Update to log a warning and
use the original expression as the diagnostic summary.

This allows creation of warnings-messages-NEXT.txt, since the message
for -Wreturn-type contains %enum_select<FalloffFunctionKind>.  A future
patch will improve the processing for %enum_select.